[arci] Change the wait logic
Before this change, if the timeout was 10 sec,
the smallest wait rate was 1 Hz.
It means that even if the sent trajectory was very short (like 10msec),
it waits 1sec at least.

By this change the unit duration is set to 10msec.
(I'm not sure that it is enough small)
It means the smallest sleep is 10 msec always, and nothing to
do with the timeout sec.